[integration tests] Move Bootstrap tests in a dedicated package
