+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/bootstrap/SecureBootstrapTest.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.integration.tests;
+package org.eclipse.leshan.integration.tests.bootstrap;
 
 import static org.eclipse.leshan.integration.tests.util.BootstrapConfigTestBuilder.givenBootstrapConfig;
 import static org.eclipse.leshan.integration.tests.util.Credentials.GOOD_PSK_ID;
